long line highlight only on long lines
Lot of work went into making the toggle work properly.
Using 'matchadd' & 'matchdelete' prevents it fouling
up other things that use match highlights.
But this requires storing the returned match group ID
with a scope which matches the 'matchadd' mechanism.

This turns out to be stored at window scope, so we use
window scope variables.

A new window, detected using WinBufEnter event, and then
testing for the window scope variable not existing,
should default to turning this highlight on.